Use Image.propTypes.source as href propType

+++ android/src/main/java/com/horcrux/svg/ImageShadowNode.java
@@ -89,9 +89,8 @@ public class ImageShadowNode extends RenderableShadowNode {
 
             if (src.hasKey("width") && src.hasKey("height")) {
                 mImageRatio = (float)src.getInt("width") / (float)src.getInt("height");
-            }
-            else {
-                mImageRatio = (float)0.0;
+            } else {
+                mImageRatio = 0f;
             }
             mUri = Uri.parse(uriString);
         }
@@ -173,7 +172,7 @@ public class ImageShadowNode extends RenderableShadowNode {
         float rectRatio = rectWidth / rectHeight;
         RectF renderRect;
 
-        if (mImageRatio == 0.0 || mImageRatio == rectRatio) {
+        if (mImageRatio == 0f || mImageRatio == rectRatio) {
             renderRect = new RectF(rect);
         } else if (mImageRatio < rectRatio) {
             renderRect = new RectF(0, 0, (int)(rectHeight * mImageRatio), (int)rectHeight);
